refactor(overrides): move PreviewOverrideValue out of :daemon:core
Preview bundles were dragging the whole 1.68 MB :daemon:core render daemon onto
every preview's runtime classpath solely to reuse one @Serializable type,
PreviewOverrideValue: catalog -> :data-preview-overrides-runtime ->
api :data-preview-overrides-core -> api :daemon:core. Of daemon:core's 708
classes a preview reaches ~18, but a dependency jar is inlined whole once any
class is reachable, so the entire daemon shipped in each bundle.

Move the type into the published, Compose/backend-free
:data-preview-overrides-core (beside PreviewOverrideDeclaration, whose default/
current fields already reference it) and invert the dependency: the daemon
protocol now depends up into overrides-core for the shared type. overrides-core
no longer depends on :daemon:core, so the daemon leaves the preview closure
entirely (0 reachable -> not inlined) rather than being sliced.

Wire format is unchanged: polymorphic serialization keys on the @SerialName
discriminators (string/int/float/bool/color), not the FQN. Pure relocation +
import updates across daemon/cli/runtime/connector; no behaviour change.

Not built locally (Gradle + ktfmt need network and time out in this
environment); relies on CI ktfmtCheckAll + module tests.
