+++ cli/src/main/kotlin/ee/schimke/composeai/cli/serve/ServeOverrides.kt
@@ -1,9 +1,9 @@
 package ee.schimke.composeai.cli.serve
 
 import ee.schimke.composeai.daemon.protocol.Orientation
-import ee.schimke.composeai.daemon.protocol.PreviewOverrideValue
 import ee.schimke.composeai.daemon.protocol.PreviewOverrides
 import ee.schimke.composeai.daemon.protocol.UiMode
+import ee.schimke.composeai.data.overrides.PreviewOverrideValue
 import java.security.MessageDigest
 
 /**
