+++ data/preview-overrides/core/src/main/kotlin/ee/schimke/composeai/data/overrides/PreviewOverrideModels.kt
@@ -1,6 +1,5 @@
 package ee.schimke.composeai.data.overrides
 
-import ee.schimke.composeai.daemon.protocol.PreviewOverrideValue
 import kotlinx.serialization.Serializable
 
 /**
